+++ core/traits/RegistryTrait.php
@@ -5,6 +5,8 @@ namespace luya\traits;
 /**
  * Registry Trait.
  *
+ * The RegistryTrait helps to handle set(), get(), has() and remove() operations for a key value based storage.
+ * 
  * Can be attached to ActiveRecords with a `name` and `value` property where name is an unique identifier.
  *
  * @author Basil Suter <basil@nadar.io>
@@ -31,30 +33,63 @@ trait RegistryTrait
     {
         return 'value';
     }
-    
+
+    private static $_data;
+
     /**
-     * Check whether a config value exists or not
+     * Loads all config data into an array.
      *
-     * @param string $name
-     * @return boolean
+     * @return array
+     * @since 1.3.0
      */
-    public static function has($name)
+    protected static function getData()
     {
-        return (self::find()->where([self::getNameAttribute() => $name])->one()) ? true : false;
+        if (self::$_data === null) {
+            self::$_data = self::find()
+                ->select([self::getValueAttribute(), self::getNameAttribute()])
+                ->indexBy(self::getNameAttribute())
+                ->column();
+        }
+
+        return self::$_data;
+    }
+
+    /**
+     * Clear data array.
+     * 
+     * @since 1.3.0
+     */
+    protected static function clearData()
+    {
+        self::$_data = null;
     }
     
     /**
-     * Get the value of a config value
+     * Check whether a config value exists or not.
+     * 
+     * If a value exists but is empty, has will return false.
      *
-     * @param string $name
-     * @return string|null
+     * @param string $name The key to lookup. If not found false is returned.
+     * @return boolean Whether the key exists or not.
+     */
+    public static function has($name)
+    {
+        return array_key_exists($name, self::getData());
+    }
+    
+    /**
+     * Get the value of a config value.
+     * 
+     * Returns the value from the registry for the given $name, if not found the defaultValue is returned.  
+     *
+     * @param string $name The key to lookup. 
+     * @param mixed $defaultValue The default value to return if the key does not exist.
+     * @return mixed
      */
     public static function get($name, $defaultValue = null)
     {
-        $model = self::find()->where([self::getNameAttribute() => $name])->asArray()->one();
-    
-        if ($model) {
-            return $model[self::getValueAttribute()];
+        if (self::has($name)) {
+            return self::getData()[$name];
         }
     
         return $defaultValue;
@@ -62,15 +97,17 @@ trait RegistryTrait
     
     /**
      * Store or Update an existing/new config value.
+     * 
+     * If the config value is not found, a new record will be created.
      *
-     * @param string $name
-     * @param string $value
-     * @return boolean
+     * @param string $name They config key
+     * @param string $value They config value. When working with array data, encode the data first with Json::encode.
+     * @return boolean Whether saving was successfull or not.
      */
     public static function set($name, $value)
     {
         $model = self::find()->where([self::getNameAttribute() => $name])->one();
-    
+        self::clearData();
         if ($model) {
             return (bool) $model->updateAttributes([
                 self::getValueAttribute() => $value,
@@ -84,16 +121,19 @@ trait RegistryTrait
     }
 
     /**
-     * Remove an existing config value
+     * Remove an existing config value.
+     * 
+     * If the value is not found in the config, false is returned.
      *
-     * @param string $name
-     * @return bool
+     * @param string $name The key to remove.
+     * @return bool If element was found and deleting was successfull true is returned, otherwise false.
      */
     public static function remove($name)
     {
         $model = self::find()->where([self::getNameAttribute() => $name])->one();
     
         if ($model) {
+            self::clearData();
             return (bool) $model->delete();
         }