+++ tests/core/web/RequestTest.php
@@ -2,6 +2,7 @@
 
 namespace luyatests\core\web;
 
+use luya\admin\Module;
 use Yii;
 use luya\web\Request;
 
@@ -9,14 +10,30 @@ class RequestTest extends \luyatests\LuyaWebTestCase
 {
     public function testisAdmin()
     {
+        $this->app->setModule('newsadmin', ['class' => Module::class]);
         $request = new Request();
         $request->forceWebRequest = true;
         $request->pathInfo = 'admin/test/';
         $this->assertEquals(true, $request->getIsAdmin(true));
+
+        $request = new Request();
+        $request->forceWebRequest = true;
+        $request->pathInfo = 'admin/';
+        $this->assertEquals(true, $request->getIsAdmin(true));
         
         $request = new Request();
         $request->forceWebRequest = true;
+        $request->pathInfo = 'admin';
+        $this->assertEquals(true, $request->getIsAdmin(true));
+
+        $request = new Request();
+        $request->forceWebRequest = true;
         $request->pathInfo = 'nothing/inside/test/';
+        $this->assertEquals(false, $request->isAdmin);
+
+        $request = new Request();
+        $request->forceWebRequest = true;
+        $request->pathInfo = 'administrator';
         $this->assertEquals(false, $request->isAdmin);
         
         $request = new Request();